requirements.txt
alpaca-py>=0.30.0
anthropic>=0.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.40
tqdm>=4.66.0
